+++ src/propertyFields/collectionData/collectionDataViewer/ICollectionDataViewerState.ts
@@ -1,0 +1,4 @@
+export interface ICollectionDataViewerState {
+  crntItems: any[];
+  inCreationItem?: any;
+}