+++ src/propertyFields/colorPicker/PropertyFieldColorPicker.ts
@@ -1,13 +1,12 @@
+import { IPropertyPaneField, PropertyPaneFieldType } from '@microsoft/sp-webpart-base';
+import { getColorFromString, IColor } from 'office-ui-fabric-react/lib/utilities/color';
 import * as React from 'react';
 import * as ReactDom from 'react-dom';
+
 import {
-  IPropertyPaneField,
-  PropertyPaneFieldType
-} from '@microsoft/sp-webpart-base';
-import { 
-	PropertyFieldColorPickerStyle,
-	IPropertyFieldColorPickerProps,
-	IPropertyFieldColorPickerPropsInternal
+    IPropertyFieldColorPickerProps,
+    IPropertyFieldColorPickerPropsInternal,
+    PropertyFieldColorPickerStyle,
 } from './IPropertyFieldColorPicker';
 import { IPropertyFieldColorPickerHostProps } from './IPropertyFieldColorPickerHost';
 import PropertyFieldColorPickerHost from './PropertyFieldColorPickerHost';
@@ -20,6 +19,7 @@ class PropertyFieldColorPickerBuilder implements IPropertyPaneField<IPropertyFie
 	public properties: IPropertyFieldColorPickerPropsInternal;
 	private elem: HTMLElement;
 	private color: string;
+	private valueAsObject: boolean;
 	private changeCB?: (targetProperty?: string, newValue?: any) => void;
 
 	public constructor(_targetProperty: string, _properties: IPropertyFieldColorPickerProps) {
@@ -36,7 +36,18 @@ class PropertyFieldColorPickerBuilder implements IPropertyPaneField<IPropertyFie
 			iconName: _properties.iconName,
 			onRender: this.onRender.bind(this)
 		};
-		this.color = _properties.selectedColor || '#ffffff';
+
+		if(typeof _properties.selectedColor === 'undefined') {
+			this.color = '#ffffff';
+		} else {
+			if(typeof _properties.selectedColor === 'string') {
+				this.color = _properties.selectedColor;
+			} else {
+				this.color = _properties.selectedColor.str;
+			}
+		}
+
+		this.valueAsObject = _properties.valueAsObject;
 	}
 
 	public render(): void {
@@ -67,14 +78,17 @@ class PropertyFieldColorPickerBuilder implements IPropertyPaneField<IPropertyFie
 
 	private onColorChanged(newColor: string): void {
 		if (this.properties.onPropertyChange && newColor !== null) {
-			this.properties.onPropertyChange(this.targetProperty, this.color, newColor);
+			let newValue: string | IColor = (this.valueAsObject ? getColorFromString(newColor) : newColor);
+			let oldValue: string | IColor = (this.valueAsObject ? getColorFromString(this.color) : this.color);
 			this.color = newColor;
-			this.properties.properties[this.targetProperty] = newColor;
+			this.properties.onPropertyChange(this.targetProperty, oldValue, newValue);
+			this.properties.properties[this.targetProperty] = newValue;
 			if (typeof this.changeCB !== 'undefined' && this.changeCB !== null) {
-				this.changeCB(this.targetProperty, newColor);
+				this.changeCB(this.targetProperty, newValue);
 			}
 		}
 	}
+
 }
 
 export function PropertyFieldColorPicker(targetProperty: string, properties: IPropertyFieldColorPickerProps): IPropertyPaneField<IPropertyFieldColorPickerProps> {